feat: 숫자 범위 검사도 추가. 2147483647 이하만 가능
Int 범위를 넘는 입력값에 대해서 오류 발생. 결과는 long 타입으로 변경.

+++ src/main/java/calculator/Application.java
@@ -5,7 +5,7 @@ public class Application {
         InputValidator inputValidator = new InputValidator();
         InputExtractor inputExtractor = new InputExtractor(inputValidator);
         Calculator calculator = Calculator.getInstance(inputExtractor);
-        int result = calculator.addInputNums(InputView.inputWords());
+        long result = calculator.addInputNums(InputView.inputWords());
         OutputView.outputResult(result);
     }
 }
