+++ src/main/java/calculator/OutputView.java
@@ -3,7 +3,7 @@ package calculator;
 public class OutputView {
     private static final String RESULT = "결과 : ";
 
-    public static void outputResult(int result) {
+    public static void outputResult(long result) {
         System.out.println(RESULT + result);
     }
 }
